move HIDDefs.h to driver/hid
